requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.3
dj-database-url==0.5.0
Django==3.0.8
django-ckeditor==5.9.0
django-js-asset==1.2.2
flake8==3.8.3
gunicorn==20.0.4
isort==4.3.21
jedi==0.17.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.7.0
Pillow==7.2.0
pluggy==0.13.1
psycopg2==2.8.5
pycodestyle==2.6.0
pydocstyle==5.0.2
pyflakes==2.2.0
pylint==2.5.3
python-decouple==3.3
python-jsonrpc-server==0.3.4
python-language-server==0.34.1
pytz==2020.1
rope==0.17.0
six==1.15.0
snowballstemmer==2.0.0
sqlparse==0.3.1
toml==0.10.1
ujson==1.35
whitenoise==5.1.0
wrapt==1.12.1
yapf==0.30.0
